Add method to describe Born term correlations
This is a bit messy but the commit includes correlations which
describe how interaction terms in the Born term of the cMoP
expansion are correlated.  There are two interaction Liouvillians
in the Born term and they must couple the same clusters to be
non-zero.  As the algorithm only propagates the effect of an
interaction on the reduced density matrix (as translational
invariance allows us to assume that neighbouring clusters have
the same density matrix), the function LBT excludes uncorrelated
terms.

To calculate the correlation matrix, we cycle through all of the
possible interactions.  If the interaction is within the cluster,
this is not included in the include_vector.  Correlations between
sites are determined and we take the tensor product of
correlation_matrix with a 4x4 matrix of ones as there are four
types of interaction (two dissipative and two unitary).