Improve method for inter-cluster interactions in optical lattices
Due to the assumption of translational invariance in the one-dimensional
optical lattice simulations, the interaction strenght for terms between
clusters entirely factorizes.  This means that instead of cycling over
interactions between all clusters, it is possible to cycle over all of the
interaction strengths to get an effective interaction strength and then
it is only necessary to consider interactions between the simulated cluster
and a single `environment' cluster.  This should speed up cluster mean-
field and cMoP calculations for optical dipole lattices considerably.